# docs/api.openapi.yaml
openapi: 3.1.0
info:
  title: SubnetIQ REST and Streaming API
  version: 1.0.0
  description:
    'Guest mathematics and lookups; Supabase bearer-token persistence; typed assistant SSE. Requests
    accept application/json and are limited to 512 KiB, except project POST/PATCH envelopes (2.25 MiB) with
    a 2 MiB plan cap. Success bodies are returned directly. Failure bodies follow Error. API responses default
    to Cache-Control: no-store. Global quota is 180 requests/minute per IP identity. Production quotas are
    atomic PostgreSQL RPC calls. No credentials means honest guest mode in development; protected operations
    return 503. All ordinary user database access retains the verified user RLS context.'
servers:
  - url: http://localhost:3001
    description: Local API
  - url: /
    description: Current deployment
...
paths:
...
  /api/calculate/{tool}:
    post:
      operationId: post_calculate_tool
      summary: Perform deterministic calculation
      tags:
        - Calculations
      security: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculationResult'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry.
        '503':
          description: Required credentials, migrations, or service are unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      description:
        Input shapes are documented in docs/implementation-contract.md and docs/calculation-semantics.md.
        The tool is calculated locally with the same pure engine used by the browser. 90 calculations per minute
        per identity.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      parameters:
        - name: tool
          in: path
          required: true
          schema:
            type: string
            enum: &id011
              - ipv4-subnet
              - ipv4-split
              - vlsm
              - aggregate
              - range-to-cidr
              - cidr-to-range
              - overlap
              - wildcard
              - convert
              - classify
              - reverse-dns
              - netmask-table
              - ipv6-subnet
              - ipv6-format
              - eui64
              - ipv6-plan
              - ipv4-map
              - bandwidth
              - mtu
              - mac
  /api/projects:
    get:
      operationId: get_projects
      summary: List own projects
      tags:
        - Projects
      security:
        - supabaseBearer: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry.
        '503':
          description: Required credentials, migrations, or service are unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing, invalid, or expired Supabase bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
        - &id002
          name: limit
          in: query
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 200
        - &id003
          name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
            maximum: 100000
    post:
      operationId: post_projects
      summary: Create a project
      tags:
        - Projects
      security:
        - supabaseBearer: []
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry.
        '503':
          description: Required credentials, migrations, or service are unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing, invalid, or expired Supabase bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProject'
      description: Project request envelope is limited to 2.25 MiB; the plan JSON is limited to 2 MiB.
...
  /api/share/{token}:
    get:
      operationId: get_share_token
      summary: Resolve a public snapshot
      tags:
        - Sharing
      security: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry.
        '503':
          description: Required credentials, migrations, or service are unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Invalid, expired, revoked, or missing link.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      description:
        Expires and revocation are checked in the service-only RPC. Responses use no-store and no-referrer.
        The same 404 is returned for invalid, expired, and revoked tokens.
      parameters:
        - name: token
          in: path
          required: true
          schema: &id004
            type: string
...
  /api/lookups/dns:
    post:
      operationId: post_lookups_dns
      summary: Query the configured public DNS resolver
      tags:
        - Lookups
      security: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculationResult'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry.
        '503':
          description: Required credentials, migrations, or service are unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '502':
          description: Resolver failure or invalid response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '504':
          description: Lookup timed out.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      description:
        PTR accepts an IP or reverse name. No arbitrary upstream URL or resolver selection. Results
        contain source, retrieval time and cache status. Rate 30 combined lookups/minute.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 253
                type:
                  type: string
                  enum:
                    - A
                    - AAAA
                    - MX
                    - TXT
                    - NS
                    - CNAME
                    - PTR
              required:
                - name
                - type
              additionalProperties: false
  /api/lookups/ip-info:
    post:
      operationId: post_lookups_ip_info
      summary: Fetch public registration information
      tags:
        - Lookups
      security: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculationResult'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry.
        '503':
          description: Required credentials, migrations, or service are unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No public registration record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '502':
          description: Registry failure or invalid response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '504':
          description: Lookup timed out.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      description:
        Single IPv4/IPv6 address only. Longest-prefix IANA bootstrap selection, fixed RIR allowlist,
        pinned public connection addresses, no redirects. Optional bounded legacy WHOIS fallback is disabled
        by default. Registration country is not geolocation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                address:
                  type: string
                  maxLength: 45
              required:
                - address
              additionalProperties: false
...
  /api/feedback:
    post:
      operationId: post_feedback
      summary: Record product feedback
      tags:
        - Accounts
      security:
        - {}
        - supabaseBearer: []
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: *id001
                  created_at:
                    type: string
                    format: date-time
                required:
                  - id
                  - created_at
                additionalProperties: false
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry.
        '503':
          description: Required credentials, migrations, or service are unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      description:
        Optional authentication. Guest feedback requires the server service key. Five submissions/hour.
        Returns success only after a real database write.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
                  minLength: 10
                  maxLength: 5000
                email:
                  type: string
                  format: email
                  maxLength: 254
              required:
                - message
              additionalProperties: false
...
components:
  securitySchemes:
    supabaseBearer:
      type: http
      scheme: bearer
      bearerFormat: Supabase access JWT
      description:
        A verified Supabase access token; API calls auth.getUser to verify it before creating a
        user-scoped database client.
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: *id004
            message: *id004
            requestId: *id004
            details: {}
          required:
            - code
            - message
          additionalProperties: false
      required:
        - error
      additionalProperties: false
    CalculationResult:
      type: object
      properties:
        toolId: *id004
        title: *id004
        normalizedInput:
          type: object
          additionalProperties: true
        summary:
          type: array
          items:
            type: object
            properties:
              label: *id004
              value:
                type:
                  - string
                  - number
                  - boolean
                  - 'null'
              description: *id004
            required:
              - label
              - value
            additionalProperties: false
        rows:
          type: array
          items:
            type: object
            additionalProperties:
              type:
                - string
                - number
                - boolean
                - 'null'
        columns:
          type: array
          items:
            type: object
            properties:
              key: *id004
              label: *id004
            required:
              - key
              - label
            additionalProperties: false
        steps:
          type: array
          items:
            type: object
            properties:
              title: *id004
              description: *id004
              formula: *id004
            required:
              - title
              - description
            additionalProperties: false
        warnings:
          type: array
          items: *id004
        blocks:
          type: array
          items:
            type: object
            additionalProperties: true
        data:
          type: object
          additionalProperties: true
        sources:
          type: array
          items:
            type: string
            format: uri
        engineVersion: *id004
      required:
        - toolId
        - title
        - normalizedInput
        - summary
        - steps
        - warnings
        - engineVersion
      additionalProperties: false
    Project:
      type: object
      properties:
        id: *id001
        name: &id005
          type: string
          minLength: 1
          maxLength: 100
        description: &id006
          type: string
          maxLength: 3000
        address_space: &id007
          type: string
          maxLength: 100
        plan: &id008
          type: object
          additionalProperties: true
        version: &id010
          type: integer
          minimum: 1
        archived: &id009
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - address_space
        - plan
        - version
        - created_at
        - updated_at
      additionalProperties: true
    CreateProject:
      type: object
      properties:
        name: *id005
        description: *id006
        address_space: *id007
        plan: *id008
        archived: *id009
      required:
        - name
      additionalProperties: false
...